Messages spacing and fitting
Made a bunch of additions to how messages are made, so that they now can be placed, and they are placed correctly compared to one another.
Also made the rect transform of content fit to the size of the messages.
Still missing functionality with scrolling. (also ignore the 30 lines of weeeeee thats fine)